+++ composeApp/src/desktopMain/kotlin/edu/dyds/movies/data/external/MoviesRemoteSource.kt
@@ -2,7 +2,9 @@ package edu.dyds.movies.data.external
 
 import edu.dyds.movies.domain.entity.Movie
 
-interface MoviesRemoteSource {
+interface PopularMoviesRemoteSource {
     suspend fun getPopularMovies(): List<Movie>
+}
+interface MovieByTitleRemoteSource{
     suspend fun getMovieByTitle(title: String): Movie
 }
